Fix seeded bugs, align training with codebase
- Fix BUG-002..006 (double accumulator, generation bump, deterministic
  solver ordering via vector_map + sorted constraints, seed XOR-fold,
  rolling-window warm-up); enable all regression tests — ctest 100% green
- Populate .specify/memory/constitution.md (was unfilled template)
- Correct constitution/AGENTS.md: std::mt19937 interop boundary, local
  build gate instead of nonexistent CI/reviewer chatmode
- Add .github/instructions/{cpp-impl,tests}.instructions.md; expand
  copilot-instructions.md
- Training doc: engine_demo naming, /speckit.* commands, Part 2 rewritten
  around pre-staged lockless-ring-buffer + fixed-string specs, fixed
  cross-refs, valid EASTL snippets, markdownlint clean
- Scrub stale Session 0x / seeded-bug references; refresh determinism
  digest matrix (verified: same seed → identical digests, all 4 scenes)

+++ include/engine_demo/sim/game_loop.h
@@ -1,10 +1,11 @@
 // Fixed-step game loop with accumulator.
 //
 // Constitutional articles satisfied:
-//   - 5 (determinism — note SEEDED DEFECT)
+//   - 5 (determinism: double accumulator)
 //   - 6 (real-time)
 //
-// SEEDED DEFECT BUG-002: see src/engine_demo/sim/game_loop.cpp.
+// FIX BUG-002: accumulator is now `double` per Article 5; the former `float`
+// accumulator drifted after ~30 s of simulated time.
 
 #pragma once
 
@@ -30,8 +31,8 @@ class [[nodiscard]] game_loop {
 
    private:
     game_loop_config m_cfg;
-    // BUG-002 anchor: this is intentionally `float` rather than `double`.
-    float m_accumulator_seconds{0.0f};
+    // Article 5: time accumulators are double, never float.
+    double m_accumulator_seconds{0.0};
 };
 
 }  // namespace engine_demo::sim
